DeletedProduct
Reestructuramos deleProduct a un Caso de uso externo, al DeleteProductController
